🔨FIX: fixes the issue with `multer` where profile picture was saved as "undefined".
This was a problem with how form-data requests stream into the server. The problem being that
the order of fields coming in is not guaranteed. The file could be processed before other text fields
(e.g. `username`) come in. This causes the `req.body.username` to be undefined which in turn causes
the user's profile picture to be saved as something like "undefined.png".

I used the "Save and Rename" approach. The file is temporarily saved and the controller handler,
renames it afterwards.